Cleaning up documentation, refactoring KPCovC
