+++ src/skmatter/utils/_pcovc_utils.py
@@ -12,24 +12,25 @@ def check_cl_fit(classifier, X, y):
             # Check compatibility with X
             fitted_classifier._validate_data(X, y, reset=False, multi_output=True)
    
-            n_classes = len(np.unique(y))
             # Check compatibility with y
             # dimension of classifier coefficients is always 2, hence we don't 
             # need to check dimension for match with Y 
             # We need to double check this...
+            n_classes = len(np.unique(y))
+
             if n_classes == 2:
                 if fitted_classifier.coef_.shape[0] != 1:
                     raise ValueError(
                         "For binary classification, expected classifier coefficients "
                         "to have shape (1, %d) but got shape %r"
-                        % (fitted_classifier.n_features_in_, fitted_classifier.coef_.shape)
+                        % (X.shape[1], fitted_classifier.coef_.shape)
                     )
             else:
                 if fitted_classifier.coef_.shape[0] != n_classes:
                     raise ValueError(
                         "For multiclass classification, expected classifier coefficients "
                         "to have shape (%d, %d) but got shape %r" 
-                        % (n_classes, fitted_classifier.n_features_in_, fitted_classifier.coef_.shape)
+                        % (n_classes, X.shape[1], fitted_classifier.coef_.shape)
                     )
                 
         except NotFittedError:
@@ -58,25 +59,20 @@ def check_svc_fit(classifier, K, X, y):
         check_is_fitted(classifier)
         fitted_classifier = deepcopy(classifier)
 
-        # Check compatibility with K
+        # Check compatibility with X
         fitted_classifier._validate_data(X, y, reset=False, multi_output=True)
         print("Pass")
-        # Check compatibility with y
-        # if fitted_regressor.dual_coef_.ndim != y.ndim:
-        #     raise ValueError(
-        #         "The regressor coefficients have a dimension incompatible "
-        #         "with the supplied target space. "
-        #         "The coefficients have dimension %d and the targets "
-        #         "have dimension %d" % (fitted_regressor.dual_coef_.ndim, y.ndim)
-        #     )
-        # elif y.ndim == 2:
-        #     if fitted_regressor.dual_coef_.shape[1] != y.shape[1]:
-        #         raise ValueError(
-        #             "The regressor coefficients have a shape incompatible "
-        #             "with the supplied target space. "
-        #             "The coefficients have shape %r and the targets "
-        #             "have shape %r" % (fitted_regressor.dual_coef_.shape, y.shape)
-        #        )
+
+        #Check compatibility with y
+        n_classes = len(np.unique(y))
+        n_sv = len(fitted_classifier.support_)
+
+        if fitted_classifier.coef_.shape[0] != n_classes - 1:
+            raise ValueError(
+                "Expected classifier coefficients "
+                "to have shape (%d, %d) but got shape %r" 
+                % (n_classes, n_sv, fitted_classifier.coef_.shape)
+            )
 
     except NotFittedError:
         fitted_classifier = clone(classifier)
