+++ tests/test_kernel_pcovc.py
@@ -27,24 +27,15 @@ class KernelPCovCBaseTest(unittest.TestCase):
 
         self.X, self.Y = get_dataset(return_X_y=True)
 
-        # # for the sake of expedience, only use a subset of the dataset
-        # idx = self.random_state.choice(len(self.X), 100)
-        # self.X = self.X[idx]
-        # self.Y = self.Y[idx]
-
-        # artificial second property
-        # self.Y = np.array(
-        #     [self.Y, self.X @ self.random_state.randint(-2, 2, (self.X.shape[-1],))]
-        # ).T
-        # self.Y = self.Y.reshape(self.X.shape[0], -1)
-
-        # self.X = SFS().fit_transform(self.X)
-        # self.Y = SFS(column_wise=True).fit_transform(self.Y)
+        # for the sake of expedience, only use a subset of the dataset
+        idx = self.random_state.choice(len(self.X), 100)
+        self.X = self.X[idx]
+        self.Y = self.Y[idx]
 
         scaler = StandardScaler()
         self.X = scaler.fit_transform(self.X)
 
-        self.model = lambda mixing=0.5, classifier=SVC(), n_components=4, **kwargs: KernelPCovC(
+        self.model = lambda mixing=0.5, classifier=LinearSVC(), n_components=4, **kwargs: KernelPCovC(
             mixing=mixing,
             classifier=classifier,
             n_components=n_components,
@@ -89,7 +80,7 @@ class KernelPCovCErrorTest(KernelPCovCBaseTest):
 
         for mixing in np.linspace(0, 1, 6):
             kpcovc = KernelPCovC(
-                mixing=mixing, n_components=4, fit_inverse_transform=True, tol=1e-12
+                mixing=mixing, n_components=4, tol=1e-12
             )
             kpcovc.fit(self.X, self.Y)
 
@@ -117,7 +108,7 @@ class KernelPCovCErrorTest(KernelPCovCBaseTest):
         for mixing in np.linspace(0, 1, 6):
             kpcovc = self.model(
                 mixing=mixing,
-                classifier=SVC(kernel="rbf", gamma=1.0),
+                classifier=LinearSVC(),
                 kernel="rbf",
                 gamma=1.0,
                 center=False,
@@ -193,7 +184,7 @@ class KernelPCovCInfrastructureTest(KernelPCovCBaseTest):
         _ = kpcovc.score(self.X, self.Y)
 
     def test_prefit_classifier(self):
-        classifier = SVC(kernel="rbf", gamma=0.1)
+        classifier = LinearSVC()
         #this fails since we are trying to call decision_function(K) on a classifier fitted with X
         #see line 340 of kernel_pcovr
         classifier.fit(self.X, self.Y)
@@ -211,7 +202,7 @@ class KernelPCovCInfrastructureTest(KernelPCovCBaseTest):
         self.assertTrue(np.allclose(W_classifier, W_kpcovc))
 
     def test_classifier_modifications(self):
-        classifier = SVC(kernel="rbf", gamma=0.1)
+        classifier = LinearSVC(kernel="rbf", gamma=0.1)
         kpcovc = self.model(mixing=0.5, classifier=classifier, kernel="rbf", gamma=0.1)
 
         # KPCovC classifier matches the original
@@ -257,7 +248,7 @@ class KernelPCovCInfrastructureTest(KernelPCovCBaseTest):
         self.assertTrue(kpcovc.classifier_ is not None)
 
     def test_incompatible_coef_shape(self):
-        classifier = SVC(kernel="linear")
+        classifier = LinearSVC()
         classifier.fit(self.X, self.Y)
         kpcovc = self.model(mixing=0.5, classifier=classifier)
 
@@ -284,10 +275,10 @@ class KernelPCovCInfrastructureTest(KernelPCovCBaseTest):
         )
 
     def test_precomputed_classification(self):
-        classifier = SVC(kernel="rbf", gamma=0.1)
+        classifier = LinearSVC()
         classifier.fit(self.X, self.Y)
         Yhat = classifier.predict(self.X)
-        W = classifier.dual_coef_.reshape(self.X.shape[0], -1)
+        W = classifier.coef_.reshape(self.X.shape[1], -1)
 
         kpcovc1 = self.model(
             mixing=0.5, classifier="precomputed", kernel="rbf", gamma=0.1, n_components=1
@@ -304,36 +295,31 @@ class KernelPCovCInfrastructureTest(KernelPCovCBaseTest):
         self.assertTrue(np.linalg.norm(t1 - t2) < self.error_tol)
 
 class KernelTests(KernelPCovCBaseTest):
-    # def test_kernel_types(self):
-    #     """Check that KernelPCovC can handle all kernels passable to sklearn
-    #     kernel classes, including callable kernels
-    #     """
+    def test_kernel_types(self):
+        """Check that KernelPCovC can handle all kernels passable to sklearn
+        kernel classes, including callable kernels
+        """
 
-    #     def _linear_kernel(X, Y):
-    #         return X @ Y.T
+        def _linear_kernel(X, Y):
+            return X @ Y.T
 
-    #     # kernel_params = {
-    #     #     "poly": {"degree": 2},
-    #     #     "rbf": {"gamma": 3.0},
-    #     #     "sigmoid": {"gamma": 3.0, "coef0": 0.5},
-    #     # }
-    #     for kernel in ["linear", "poly", "rbf", "sigmoid", "cosine", _linear_kernel]:
-    #         with self.subTest(kernel=kernel):
-    #             kpcovc = KernelPCovC(
-    #                 mixing=0.5,
-    #                 n_components=2,
-    #                 classifier=SVC(
-    #                     kernel=kernel, 
-    #                     degree=2,
-    #                     gamma=3.0,
-    #                     coef0=0.5
-    #                 ),
-    #                 kernel=kernel,
-    #                 degree=2,
-    #                 gamma=3.0,
-    #                 coef0=0.5
-    #             )
-    #             kpcovc.fit(self.X, self.Y)
+        # kernel_params = {
+        #     "poly": {"degree": 2},
+        #     "rbf": {"gamma": 3.0},
+        #     "sigmoid": {"gamma": 3.0, "coef0": 0.5},
+        # }
+        for kernel in ["linear", "poly", "rbf", "sigmoid", "cosine", _linear_kernel]:
+            with self.subTest(kernel=kernel):
+                kpcovc = KernelPCovC(
+                    mixing=0.5,
+                    n_components=2,
+                    classifier=LinearSVC(),
+                    kernel=kernel,
+                    degree=2,
+                    gamma=3.0,
+                    coef0=0.5
+                )
+                kpcovc.fit(self.X, self.Y)
 
     def test_linear_matches_pcovc(self):
         """Check that KernelPCovC returns the same results as PCovC when using a linear
@@ -351,10 +337,9 @@ class KernelTests(KernelPCovCBaseTest):
         # computing projection and predicton loss with linear KernelPCovC
         # and use the alpha from RidgeCV for level regression comparisons
         kpcovc = KernelPCovC(
-            classifier=SVC(kernel="linear", gamma='scale', coef0=0),
+            classifier=LinearSVC(),
             kernel="linear",
             gamma='scale',
-            fit_inverse_transform=True,
             **hypers,
         )
         kpcovc.fit(self.X, self.Y)
@@ -393,102 +378,45 @@ class KernelTests(KernelPCovCBaseTest):
             round(lk_ref, rounding),
         )
 
-        # """Check that KernelPCovR returns the same results as PCovR when using a linear
-        # kernel.
-        # """
-        # svc = SVC()
-        # svc.fit(self.X, self.Y)
-
-        # # common instantiation parameters for the two models
-        # hypers = dict(
-        #     mixing=0.5,
-        #     n_components=1,
-        # )
-
-        # # computing projection and predicton loss with linear KernelPCovR
-        # # and use the alpha from RidgeCV for level regression comparisons
-        # kpcovc = KernelPCovC(
-        #     classifier=SVC(kernel="linear"),
-        #     kernel="linear",
-        #     fit_inverse_transform=True,
-        #     **hypers,
-        # )
-        # kpcovr.fit(self.X, self.Y)
-        # ly = (
-        #     np.linalg.norm(self.Y - kpcovr.predict(self.X)) ** 2.0
-        #     / np.linalg.norm(self.Y) ** 2.0
-        # )
-
-        # # computing projection and predicton loss with PCovR
-        # ref_pcovr = PCovR(**hypers, regressor=ridge, space="sample")
-        # ref_pcovr.fit(self.X, self.Y)
-        # ly_ref = (
-        #     np.linalg.norm(self.Y - ref_pcovr.predict(self.X)) ** 2.0
-        #     / np.linalg.norm(self.Y) ** 2.0
-        # )
-
-        # t_ref = ref_pcovr.transform(self.X)
-        # t = kpcovr.transform(self.X)
-
-        # K = kpcovr._get_kernel(self.X)
-
-        # k_ref = t_ref @ t_ref.T
-        # k = t @ t.T
-
-        # lk_ref = np.linalg.norm(K - k_ref) ** 2.0 / np.linalg.norm(K) ** 2.0
-        # lk = np.linalg.norm(K - k) ** 2.0 / np.linalg.norm(K) ** 2.0
-
-        # rounding = 3
-        # self.assertEqual(
-        #     round(ly, rounding),
-        #     round(ly_ref, rounding),
-        # )
-
-        # self.assertEqual(
-        #     round(lk, rounding),
-        #     round(lk_ref, rounding),
-        # )
-
-
 class KernelPCovCTestSVDSolvers(KernelPCovCBaseTest):
-    # def test_svd_solvers(self):
-    #     """
-    #     Check that KPCovC works with all svd_solver modes and assigns
-    #     the right n_components
-    #     """
-    #     for solver in ["arpack", "full", "randomized", "auto"]:
-    #         with self.subTest(solver=solver):
-    #             kpcovc = self.model(tol=1e-12, n_components=None, svd_solver=solver)
-    #             kpcovc.fit(self.X, self.Y)
+    def test_svd_solvers(self):
+        """
+        Check that KPCovC works with all svd_solver modes and assigns
+        the right n_components
+        """
+        for solver in ["arpack", "full", "randomized", "auto"]:
+            with self.subTest(solver=solver):
+                kpcovc = self.model(tol=1e-12, n_components=None, svd_solver=solver)
+                kpcovc.fit(self.X, self.Y)
 
-    #             if solver == "arpack":
-    #                 self.assertTrue(kpcovc.n_components_ == self.X.shape[0] - 1)
-    #             else:
-    #                 self.assertTrue(kpcovc.n_components_ == self.X.shape[0])
+                if solver == "arpack":
+                    self.assertTrue(kpcovc.n_components_ == self.X.shape[0] - 1)
+                else:
+                    self.assertTrue(kpcovc.n_components_ == self.X.shape[0])
 
-    #     n_component_solvers = {
-    #         "mle": "full",
-    #         int(0.75 * max(self.X.shape)): "randomized",
-    #         0.1: "full",
-    #     }
-    #     for n_components, solver in n_component_solvers.items():
-    #         with self.subTest(solver=solver, n_components=n_components):
-    #             kpcovc = self.model(
-    #                 tol=1e-12, n_components=n_components, svd_solver="auto"
-    #             )
-    #             if solver == "randomized":
-    #                 n_copies = (501 // max(self.X.shape)) + 1
-    #                 X = np.hstack(np.repeat(self.X.copy(), n_copies)).reshape(
-    #                     self.X.shape[0] * n_copies, -1
-    #                 )
-    #                 Y = np.hstack(np.repeat(self.Y.copy(), n_copies)).reshape(
-    #                     self.X.shape[0] * n_copies, -1
-    #                 )
-    #                 kpcovc.fit(X, Y)
-    #             else:
-    #                 kpcovc.fit(self.X, self.Y)
+        n_component_solvers = {
+            "mle": "full",
+            int(0.75 * max(self.X.shape)): "randomized",
+            0.1: "full",
+        }
+        for n_components, solver in n_component_solvers.items():
+            with self.subTest(solver=solver, n_components=n_components):
+                kpcovc = self.model(
+                    tol=1e-12, n_components=n_components, svd_solver="auto"
+                )
+                if solver == "randomized":
+                    n_copies = (501 // max(self.X.shape)) + 1
+                    X = np.hstack(np.repeat(self.X.copy(), n_copies)).reshape(
+                        self.X.shape[0] * n_copies, -1
+                    )
+                    Y = np.hstack(np.repeat(self.Y.copy(), n_copies)).reshape(
+                        self.X.shape[0] * n_copies, -1
+                    )
+                    kpcovc.fit(X, Y)
+                else:
+                    kpcovc.fit(self.X, self.Y)
 
-    #             self.assertTrue(kpcovc._fit_svd_solver == solver)
+                self.assertTrue(kpcovc._fit_svd_solver == solver)
 
     def test_bad_solver(self):
         """
@@ -501,20 +429,20 @@ class KernelPCovCTestSVDSolvers(KernelPCovCBaseTest):
 
         self.assertTrue(str(cm.exception), "Unrecognized svd_solver='bad'" "")
 
-    # def test_good_n_components(self):
-    #     """Check that KPCovC will work with any allowed values of n_components."""
-    #     # this one should pass
-    #     kpcovc = self.model(n_components=0.5, svd_solver="full")
-    #     kpcovc.fit(self.X, self.Y)
+    def test_good_n_components(self):
+        """Check that KPCovC will work with any allowed values of n_components."""
+        # this one should pass
+        kpcovc = self.model(n_components=0.5, svd_solver="full")
+        kpcovc.fit(self.X, self.Y)
 
-    #     for svd_solver in ["auto", "full"]:
-    #         # this one should pass
-    #         kpcovc = self.model(n_components=2, svd_solver=svd_solver)
-    #         kpcovc.fit(self.X, self.Y)
+        for svd_solver in ["auto", "full"]:
+            # this one should pass
+            kpcovc = self.model(n_components=2, svd_solver=svd_solver)
+            kpcovc.fit(self.X, self.Y)
 
-    #         # this one should pass
-    #         kpcovc = self.model(n_components="mle", svd_solver=svd_solver)
-    #         kpcovc.fit(self.X, self.Y)
+            # this one should pass
+            kpcovc = self.model(n_components="mle", svd_solver=svd_solver)
+            kpcovc.fit(self.X, self.Y)
 
     def test_bad_n_components(self):
         """Check that KPCovC will not work with any prohibited values of n_components."""
